update oracle spacewalk mock adapter
