+++ pallets/oracle/src/mock.rs
@@ -27,7 +27,7 @@ use crate::{
 	self as oracle,
 	dia::DiaOracleAdapter,
 	oracle_mock::{
-		DataCollector, MockConvertMoment, MockConvertPrice, MockDiaOracle, MockOracleKeyConvertor,
+		MockConvertMoment, MockConvertPrice, MockOracleKeyConvertor, Data,
 	},
 	Config, Error, OracleKeys,
 };
@@ -229,3 +229,73 @@ where
 		// COINS.write().unwrap().clear();
 	});
 }
+
+
+
+
+thread_local! {
+	static COINS: RefCell<Vec<Data>> = RefCell::new(vec![]);
+}
+
+pub struct MockDiaOracle;
+impl DiaOracle for MockDiaOracle {
+	fn get_coin_info(
+		blockchain: Vec<u8>,
+		symbol: Vec<u8>,
+	) -> Result<dia_oracle::CoinInfo, sp_runtime::DispatchError> {
+		let key = (blockchain, symbol);
+		let mut result: Option<Data> = None;
+		COINS.with(|c| {
+			let r = c.borrow();
+			for i in &*r {
+				if i.key == key.clone() {
+					result = Some(i.clone());
+					break
+				}
+			}
+		});
+		let Some(result) = result else {
+			return Err(sp_runtime::DispatchError::Other(""));
+		};
+		let mut coin_info = CoinInfo::default();
+		coin_info.price = result.price;
+		coin_info.last_update_timestamp = result.timestamp;
+
+		Ok(coin_info)
+	}
+
+	fn get_value(
+		blockchain: Vec<u8>,
+		symbol: Vec<u8>,
+	) -> Result<dia_oracle::PriceInfo, sp_runtime::DispatchError> {
+		todo!()
+	}
+}
+
+pub struct DataCollector;
+impl DataProvider<Key, TimestampedValue<UnsignedFixedPoint, Moment>> for DataCollector {
+	fn get(key: &Key) -> Option<TimestampedValue<UnsignedFixedPoint, Moment>> {
+		todo!()
+	}
+}
+impl orml_oracle::DataFeeder<Key, TimestampedValue<UnsignedFixedPoint, Moment>, AccountId>
+	for DataCollector
+{
+	fn feed_value(
+		who: AccountId,
+		key: Key,
+		value: TimestampedValue<UnsignedFixedPoint, Moment>,
+	) -> sp_runtime::DispatchResult {
+		let key = MockOracleKeyConvertor::convert(key).unwrap();
+		let r = value.value.into_inner();
+
+		let data = Data { key, price: r, timestamp: value.timestamp };
+
+		COINS.with(|coins| {
+			let mut r = coins.borrow_mut();
+			r.push(data)
+		});
+		Ok(())
+	}
+}
+
